Move user authentication to seperate file
